+++ src/main/java/net/zhaiji/chestcavitybeyond/api/function/OtherOrganChangeConsumer.java
@@ -11,8 +11,8 @@ public interface OtherOrganChangeConsumer {
     /**
      * @param context       当前器官（接收回调的器官）的上下文
      * @param changedIndex  变化的器官槽位索引
-     * @param changedStack  变化的器官物品栈
-     * @param isAdded       true=添加，false=移除
+     * @param oldStack      旧器官物品栈（可能为空）
+     * @param newStack      新器官物品栈（可能为空）
      */
-    void accept(ChestCavitySlotContext context, int changedIndex, ItemStack changedStack, boolean isAdded);
+    void accept(ChestCavitySlotContext context, int changedIndex, ItemStack oldStack, ItemStack newStack);
 }